Extract workspace input helpers
